Converted to multi module project
The project is now a multi module project.  One module consists of the
batch job code as well as the configuration to build and publish the
docker image.  The other module will contain the code for a simple web
application that will illustrate orchestrating Spring Batch jobs as
Lattice Tasks.